✨ Feat: Tribuo 기반 카드 순위화(LTR) 베이스 모델 구축
- Tribuo 라이브러리를 이용한 카드 추천용 회귀 모델 학습 파이프라인 구현
- historical_transactions 데이터를 활용한 기초 피처(금액, 건수, 승인율) 추출 로직 추가
- 학습된 모델 저장(.gdpc) 및 실시간 점수 추론을 위한 InferenceService 구현
- 추천 후보군에 대한 점수 산출 및 내림차순 정렬 서비스(RankingRecommendationService) 추가

Closes #17
